Upgrade Lucene 6.6.1 -> 7.0.1
Switch to Lucene 7.x DocValues iterator model for constructing security filter. Collapse container ID and resource ID into a single "SecurityContext" field to eliminate unnecessary (and seldom needed) lookup. Both changes should result in big security filter performance improvements.

+++ search/src/org/labkey/search/SearchController.java
@@ -556,7 +556,7 @@ public class SearchController extends SpringActionController
             final String query = form.getQueryString();
             final JSONObject response = new JSONObject();
             Object[] arr = new Object[0];
-            int totalHits = 0;
+            long totalHits = 0;
 
             if (null != StringUtils.trimToNull(query))
             {
